Database
Managed to create a form that uploads and gets data from the database. I also reworked the home page loop to use the data from the database. I am playing around with custom keys but if that doesnt work then I can live with ugly urls as I believe that is the main negative of my current system. I dont think security is an issue.