Working Edit Button
Ok, so it's literally less secure than a low security prison with a train station but dear lord I actually managed to get it working. Right now you can see the current article text and are able to change it using the text box. Obviously it needs to be more secure but I'm really excited I was able to get something working.