Changing Title + Editing Different Articles
So It turns out I didnt properly allow for the editing for multiple articles. I had previously hardcoded the article to be edited which meant I was only changing Blog 1 regardless of the page I was on. I fixed that and added the ability to change an articles title on page.