Edit Form Render
Good progress on restricting the edit form to the correct user. It's currently broken when nobody is logged in but when it works it only renders the edit form when the logged in user email matches the articles authors email in the database. Not a perfect solution but it will do for my needs.