Stop Gap Solution Found
The issue is with the firebase generated keys. Previously I was using the database id number as a match for routing but with the switch random letters and numbers it's not as simple. For the moment when submitting an article you're sent to the home page. Not a big deal as the latest article is at the top of the page. Will need to work on this.