Keyboard Control
Want to implement keyboard control. I can have the myKnews page work with Control + Enter but I i also want to to se the same component to work with multiple forms. I think it's possible I just need to get a bit more refreshed with the workings of react. This is a good example of me probably needing  to do refreshers before jumping back into a full project.... but where's the fun in that.