Working Lazy Loading
An honest to god working implementation of Lazy Loading. The solution was frustratingly simple in the end, simply concatenating a new array of articles onto the existing articles.  Annoyingly I'd already done most of the work in setting up the states and incrementing through the array, I just hadnt connected the dots. There are still things to be sorted out like what happens when you click on an article and try to return home but I'm really happy with the progress here.