Version 8
Ok, so instead of just deploying I basically overhauled the look of the articles, removed a bunch of features and slimmed down the scale of the proejct. Read the blog for all the details but basically I'm a monster and future me has a heck of a job figuring out the branches.