Controlling Recommended Reading Bar Items
Right now the area is populated with the first 4 articles in order of key. I want to be able to specify whatever article I want directly from the database.
My solution is to bascally copy and paste the first main query to the database and then appended the results to the state. It is working as intended but i think there is definitley a better way of doing this.