Version 9 Release
I want to get something on the website so here is theKnews V9

Added in Redesigned Back Button to article view...
...then quickly replaced it with a Home button
Added Previous Article button to article view
Added Next Article button to article view
These buttons are quite janky as they will only work as long as the next/prev page id is numerically adjacent(1>2>3 will work, 1>4>8 will not). I wanted to have the basic functionality in place as the page id issue is a larger issue I need to address.
Still, it was good to get this working relatively simply. As already said, I need to rethink the way I bring things in from the database as when I start deleting articles the current system won't work
Added Tags to articles that will let you filter articles based on type of content
Added Tagbar that is currently a defacto website navigation menu
This currently only works on the home page, so you need to go back to the home page to notice the difference.
I could have left this out of the release but I wanted it here as a demo of what it will look like when complete. It's good to see progress even when not fully functional.
Improved appearance of recommended reading area
Tweaked appearance of news squares on home page