Ability to Hide Articles from inside articles - V0.4
Solved the issue by using the the hide article component in the article page as the hide button on the home page. There's more work to be done in eliminating code bloat as the new component was using similiar code to the old hide button. However, it's good progress and I'm happy with what i have achieved.