Moving Scroll to it's own component...or trying to at least. - V0.1
In an effort to reduce the size and complexity of the tags file, i am trying to separate the scroll function into it's own compoment. This would also help reduce the file size for the home page as it uses the exact same piece of code. So far I've made good progress but Im struggling with getting the main array into the correct format when entering and leaving state.