Creating new component for rendering Cards - V0.1
Why the hell not? This is going to open up so many cans of worms and issues but If I want to get the codebase to be more efficent and slimmer, then it needs to happen. It also centralises the view for the cards across the home page and tag pages.