Creating new component for rendering Cards - V0.2
Having now created a component for rendering the cards I realise that new news-item-loop-view is bascially already that. Sure it's still large but by the time I've moved everything out of the the original file, there's bascially about 5 lines left in the file. I know the point was cut down on bloat but I actually think i'm adding bloat. The news-item-loop-view can be used across the website as is with minmal problems.