Adding Card Size Controls To Tag Page - v1
Added the card controls component to the tags page.
Currently getting an error when selecting an option and reloading the page or clicking away. Not sure what the issue is yet but disabling the scrollCheck component let's the tags page load with functioning card controls.  Suspect it has to do with articlesArray coming from the tags file.