Handling All Articles Being Hidden - V1
Just getting started, not a lot to report. My main idea is to have a method where the website can handle all articles being hidden. Currently, you get a blank page but I want change this to ensure that there is always an article on page if available. At the moment the scrollCheck component won't fire, leaving an empty page.