Handling All Articles Being Hidden - V4 - Actual Progress
I managed to create a brand new filtered array. It's not the end but with this array I can ensure a minimum number of articles loaded at the initial load, regardless of how many articles have been hidden. That's the idea, next I need to hook up the hide article button to change the value of the array when the article is hidden.