Handling All Articles Being Hidden - V7
You can now press the hide article button and an article replaces the hidden article without causing duplication issues. It's not working correctly as the article is replaces it is not the the first item from the array. My code has also taken a turn for the messy so that needs to be cleaned up.