Handling All Articles Being Hidden - V9.1 - scrollCheck V2.2
Woo, managed to make some progress. In the hide-article component I added another filter to filter the leftoverarticles array for the id of the card that was hidden. It's the same as the filter for the initial load. I should look into just using a single fillter as it's the same code just using different arrays as reference.