Handling All Articles Being Hidden - V9.2 - Redoing the Tags Page
So yes. I have decided that, after all of the work with scroll check that it would be a good idea to redo the tags component using the same thinking as the new method of loading articles introduced with the new scroll check component. This will take away the need for multiple identical database calls and reduce code bloat. I honestly think this is a good idea and will make for a more streamlined and easier to manage code base.