Handling All Articles Being Hidden - V9.2.1 - Up and Running
I have removed the old code and implemented a filter to get an array of the correct to the main render method. It's still early but other than a few issues with the scroll check component, things are looking good. Clicking on a tag gets you a page full of the tag you clicked on.