Handling All Articles Being Hidden - V9.2.2 - Minor Progress
Made some progress in allowing the tags page to function with the hide button. However, I've noticed a problem that needs to be solved before I can move on. Basically, I am going to remove all of the database calls from the website and have a single call higher up the render chain that can be shared across the entire website without the need to use local storage.