Handling All Articles Being Hidden - V9.2.3 - Kill Me.
the tags pages arent changing without a reload and I have no earthly idea why.seriously. what the hell is going on here. I'm getting worked up so this commit is my last for the week. On monday i'm going to start a fresh. Redo the entire functionality from top to bottom. Rip out the tags bar and start the entire thing again.