Handling All Articles Being Hidden - V9.2.15.1 - Figured out props to article
I figured out how I'm passing props from the main page to the article page. I also came to the conclusion that if I want direct to page links to work then I need to query the database in news-page.js. Unless I want to call the entire database when the entire website loads, it's probably better to just be doing  a single call for each article. i do need to clean up the process but it probably needs to remain as is.