Handling All Articles Being Hidden - V9.2.17 - Converting .map into compoent
A side project while I work on tags, I'm (once again) seeing If I can take the .map, that outputs the article cards, and convert it into a reusable component. My initial test looks positive and if i can get this working it will save a lot of space and be a significant optimisation by having a single card renderer component to work with instead of about 5 as I currently have.