Handling All Articles Being Hidden - V9.2.19 - Hiding on "tag" page
I have added the ability to hide articles on the "author/post date" specific page( i don't know what to call this other than the tag page even though they're not tags"). The rest of the website respects the hide but reloading the page will display articles that should be hidden.