Handling All Articles Being Hidden - V9.2.19.1
The hide button now works as normal on the tags page. Reloading the page will not render articles that should be hidden. There are some weird issues though that I need to deal with. These are that sometimes nothing appears on certain pages and random articles being added to the home page after hiding articles on tags.