Handling All Articles Being Hidden - V9.2.24 - Combining Search / Tags component
I'm trying to combine the functionality of the 2 tags components. One can handle search via url and the other directly through the article page. I want to combine the functionality of both into one component as there's no real reason to have separate components that do the same thing with the only difference being what parameters are used to query the database and how that query is constructed.