Handling All Articles Being Hidden - V9.2.25 - Fixed issue when hiding certain articles
Fixed an issue where hiding an article on the dedicated tags and author pages would result in an empty page when navigating to the home page. The issue was solved by placing a second database call in tags3.1 which ensured that a full database array was retrieved in order for the hide button to properly work.