Handling All Articles Being Hidden - V9.3 - Integrating Swipe into new Hide Component
Fixed swipe to hide in order to work with the new hide component...by also creating a new hide component. The new hide component is purely functional so there's no need for it to be a normal component. I will need to move the UI for button press over to the render card component but thats not a big deal. I have also removed the checkCache component from the renderCard component as it has become obsolete with the development of the new hideArticle component. Component.