Handling All Articles Being Hidden - V9.3.1 - Move to new functional hide article component
Moved over to the new functional hide article component. It's the same code just converted to a functional component as it doesn't use state. However, I have encountered an old issue relating to hiding articles and so i am going to attempt a fix. The issue is that hiding on a tags page returns a blank page when going to the home page.