Handling All Articles Being Hidden - V9.5.5 - speedKnews
Instead of overhauling the article pages, i have instead settled on redoing the popup as "speedKnews". An area of the website i imagine would be used for quickly reading through the latest news items. speedKnews is designed to be lightweight with minimal content(no tags, social stuff recommended reading etc) so make's sense to be used with the next/prev buttons.