Handling All Articles Being Hidden - V9.14.1 - Tweaks and a new component
My version naming structure is just all over the place but surprisingly I do understand it.
Anyway, I've started work on a message that appears when you hide an article. I've already created the UI for it and tested it with the hide button on home. The next step is to create a separate component so that i can be used wherever you can hide an article. It's currently just sitting in the news-item-loop-view file.