Handling All Articles Being Hidden - V9.15 - Fixed issue with recommended reading
Had an issue where clicking on a recommended reading link would take you to a blank page. The issue here was that I never updated the URL in the recommended reading component when I added /home/articles to the url for news pages. Making this change corrected the problem although this component hasn't been touched for a very long time so maybe I need to devote some time to this at some point.