12.1.6 - Fixed issue with Full Website Reset Button
When pressing the full website reset button you would be met with an error. This is because I was clearing the local cache but not replacing the card size with anything which was returning null instead. Setting a value for card size immediately after clearing the cache fixes the issue.