12.1.8 - Tweaking menu to reload on the same page
I added some code to be able to reload the page on the same page you reset card size and unhide articles. This is useful for tags where you might not want to return to the home page after unhiding articles or changing card size. The code works, reloading the page on the same page throwing up an error when attempting to hide articles.