12.3.2 - Created Component
Created a createBookmark component that adds an ID to a bookmark array. This is stored in local storage and will be used to output the list of bookmarks on its own page. Also, after adding a bookmark you get a border around the card , however this does not currently survive page reload but it's a start.