12.3.4 - Created Bookmark page with saved bookmarks
I have managed to create a page that will display cards bookmarked. It does the job but I think I can implement this better. At a minimum I should be able to unhide bookmarks without clearing anything else and the current code doesn't allow this. However, it's a good start and I don't think getting this to a point I am happy with will be too much of a hassle.