12.3.7 - More work on Visual ID of bookmarks
You can now bookmark a page by swiping. When bookmarked you get a border around the card itself(not final) and the card is added to the bookmarks page. When you return home, the card will still have a border around the card. This can be removed   using the options menu.