12.3.7.1
Not a lot to say here, im trying to find a solution to a bug that shows after reloading a page with a hidden bookmarked article. Currently getting an error and I'm not sure how to fix it. I know that the issue is with trying to add a class name to an element not in the dom. I just don't know at this point how to get around it.