12.3.9 - Added mark as read and hide button
Added a button that lets you mark and article as read and also hide it from view in the bookmarks page. It's a bit awkward to test as there are a number of different combinations of hiding, unhiding and refreshing that need to be accounted for. I've fixed a few things so far but will need to be on the look out. I also still need to remove the card from bookmarks after pressing the mark as read and hide button as it will reappear on reload in the bookmarks page.