12.3.13.2 - A bit of a tangent
Im havig issues with creating an interactive ui for the on card bookmark controls and I think switching from a stateless functional component over to a stateful component might help me out and make things a bit more straightfoward. It might not and I might need to restructure my code but what's the harm in giving it a crack.