12.3.13.5 - Fixed issue with bookmark toggle
Manged to fix the issue with toggling bookmarks by adding a set.item to the remove bookmark component as well as passing the articles bookmark status through props and setting that value as the initial bookmark state. However, the clear bookmark buttons are now returning null. Still, i can toggle bookmark with a single button so that's a good breakthrough.