12.3.13.6 - Fixed issue with clear bookmarks button
I've added a button that let's you toggle a bookmark. It works but every now and then you need to press the button twice to create the bookmark. I'm unable to replicate the issue but it does happen. It may be an issue with the area you need to click and im just not pressing in the right area but I will need to keep an eye out for this. I've fixed most of the issue that causes the website to break but the double click issue is something I will need to watch out for.