12.3.14 - Changes to Bookmark page controls
Switched from the old bookmark controls to be the same as the cards on the home page. I have run into an issue with hiding articles and their behaviour with bookmarks. Hiding the articles is fine, they will remove from bookmarks but clearing ALL bookmarks doesn't do anything and they stay in the bookmarks page.