12.3.17 - Added Unhide All Articles button
I also changed the unhide articles button in the options menu by creating a new component for both the bookmarks page and the options menu. This might have caused issues with other parts of the website but this is the right move and i'll deal with those problems if they show up.