12.4.3.1 - Moved Filter Views to separate component
This one is a bit more complicated than liteKnews. liteKnews was pretty self-isolation and only really needed a button detection to the parent. filters interact with the main card render so more needs to be passed. I've also tied the filter option views directly into the core of how the cards are rendered. I can get this done but it's going to be a delicate process.