12.4.3.2 - Some progress
I can now select a filter and the page doesn't crash...so that's not nothing. In truth, i'm almost there. I can select a filter and pass the correct array to the main render view, i just need to update the parent component when clicking one of the filter options. Right now I need to reload the page to see the change.