V12.5.5 - Added onCard controls to articles and more
Noticing that I was redoing work I've already done in creating a new component for bookmarking inside an article, i decided to instead use the existing onCardControls component. However, the "new" component i created didn't go to waste as I've incorporated the method for toggling the bookmark icon on/off over to the onCardControls component. I still need to integrate the mark as read functionality and hide article button but most of the heavy lifting is done.