V12.5.6 - Article bookmark controls progress.
I removed the mark as read button from the article bookmark controls as I want this to be an automatic thing
I also integrated the hide article button here although this has the potential to create some weird interactions. For example you can hide an article on the page, leave that page and return to it using the browsers back functionality. It doesn't break anything so it's not a problem but you can then bookmark that article and it will return to the home screen as a bookmark. Again not breaking anything just weird and something i will need to think about.