V12.5.8 - Adding Bookmarks to LiteKnews
Integrating bookmarks into liteKnews has proven trickier than i thought it would be. I can get the fundamential create a bookmark stuff working but that UI is harder to update due to the way I have designed the entire component. So, in an attempt to fix it I am going to work on a modfied version of liteKnews. I don't want to go down a rabbit hole here so it should be a fairly simple experiment to see if a map would be better than what I currently have.