V12.5.8.1 - Progress...maybe
I created a new component for everything on liteKnews except the controls and It looks like it's working. The main thing I have gained here is state unique to each "page" in liteKnews which helps me send different props to the cardControls component. .I Haven't tested in depth yet, but it looks promising