V12.5.11 - Tweaks to Hiding Bookmark
I have tweaked hiding bookmarks so that you can now no longer hide a totally hide a bookmarked article except on the bookmark page. This means you can hide a bookmarked article from the home page but it will still be viewable on the bookmarks page. In order to full remove this article you then need to unbookmark it or press the new perma hide button.