V12.5.12 - Integrating ScrollCheck into bookmarks
I have partially integrated the infinite scroll functionality with bookmarks and mark as read. Article's marked as read will display that like the rest of the website. However, bookmarking a newly loaded article will still bookmark it and be viewable in the bookmark page, but the on card UI will not show this. This is a bug related to how I split up the initially loaded articles from the articles to be loaded in. I intend to redo both the scrollCheck and hide article components in a future release to fix this. For now, i'm content with the current situation and will note this is a bug to be fixed.