V12.6.1 - Started on new component, some progress
I started from a basic point of grabbing the current like, changing it and returning it to the array to be reflected on the page. After some minor tweaks to markAsRead which i set up to call a fresh database everytime it mounted, I managed to get a simple counter working. This is a trivial piece of progress but it means the UI is in place, I now just need to work on the database side of things.