V12.6.3.1 - Small Steps
I'm at a point with the like button where I can properly update the array and the value of like is what I want. This is insanely annoying and I don't know why i am finding this so troublesome. It's a glorified counter and Im struggling with it. After everything else I've done this is the thing that's getting to me. Enough for today, back tommorow to see if I can do better.