V12.6.8.3 - Close...
I've said this before but I am almost there. I have the UI behaving as I want for the most part, the main issue is that there if you click dislike and then like, the like counter is increased even if it has already been increased. I am looking for a solution where the like counter isn't increased if it has been increased.