V12.6.9.1 - Is this is?
Using everything I have learned since starting this rework, I have put together the exact  UI and button switcing system that I want. The code is a simple set of if statements that  changes liked and counter status depending on the button. The next step is to integrate the buttons with the main website array and database and that should be it.