V12.6.9.3 - Database Integrated
The like/dislike buttons are now integrated into the database. The flaw is that you can clear cache and vote to add more votes. however, for this project that's not a big deal. The next step is to try and merge some (or all) of the like/dislike buttons code into 1. There's a lot of duplicate code here so maybe something can be shortended.