12.8.X.1 - Well, i did it
So i have my new method of getting articles from the database. It allows me to access the same article directly and via url. The next step is to see how smoothly I can integrate this into the current article setup. So far, nothing seems to be affected(and it shouldnt, this change should only change articles) but I'll need to see.