12.8.X.2 - Experiment Successfull
It looks like this is the way to go. I have basically switched from getting stuff from the database using the key, to using the Id instead. This is much more flexible and easier to integrate with other parts of the website. I am going to keep testing to see if any issues come up but so far it seems promising. I also removed the images temporarily because they're beginning to use a lot of my hosting data.