12.10.1.8 - Added bookmark button
The bookmark button is now in. Not a lot to say here, it's just a simple copy and paste job. It's the other stuff that might be more troublesome as I try to wrangle it into the new menu. I've also added a touch of colour on hover but react.bootstrap is making it tricky to get right due to its in built styling. Nothing that can't be overcome with css but there's always a fight.