12.10.1.10.1 - Integrated filter functionality
Turns out the filter view is required to be called on render to populate the page with articles. So removing the current component and putting it in a dropdown menu causes the page to be blank on render until activating the dropdown. I have hide the old filter buttons but i will come back around to fixing this later(13.4). The next step at the moment is to work on the design of the filer dropdown menu