12.10.1.11.1 - Further Style tweaking and minor restructure.
I've done a bit of style unification here. Previously, the components for tags, card size and options had different layouts and levels of div to then. Now I have cut back and simplified each of these components so that they have a similar structure, making styling easier as its one set of styles for all dropdowns. As a result the options menu has also been integrated and is working as expected except for the clear bookmarks button which is not currently working.