12.10.1.15.1 - Component Created
Converting the UI bar in the card area into a nav bar for the entire website is the logical next step for the project. It will solve in advance issues with filtering, accessing bookmarks and options menu among other things. Whether I can get this working is the fun question i hope to answer.