12.10.1.15.9 - Working on Bookmark UI Bar
I've got 3 of the 4 buttons in place with a new UI specific to the bookmark page. The button to be added is a new mark all as read button which shouldn't be too much of an issue to create as it should be the opposite of the mark all unread.