12.10.1.15.11 - null card size bug fix
Fixed an issue where a browser with no previous history of accessing theKnews would throw a null error. This was due to the news-item-loop-view component requiring a call to local storage to function. I created a check for null, passing fall back values to account for this and the issue resolved.