12.10.1.22 - Added navbar to liteKnews
ALready added the bookmark link and controls as well as the link to the full article. I am going to try adding the article controls to the nav bar as well. This might not be what I want in either looks or usability, but I do want to give it a try and see how it looks. I will probably not like it but it can't hurt to try out.