12.10.1.22.1 - Added liteKnews controls to navbar, making plans
I've added the liteKnews controls to the navbar as planned. Next I am going to try pinning the navbar to the bottom of the page for mobile. On larger devices it doesn't matter as much, but for smaller devices it makes sense for controls to be close to your hand.