12.10.1.26 - Experimenting with page titles
I've been trying to see if I can create a way of quickly showing where you are in the website at all. Times' I tried multiple things like a title in the header next to theKnews and gigantic bold text at the top of the page. I was never fully happy with either solution as I think they both disrupt the overall design. Still, it was a good experiment and I might incorporate some of these ideas later on as I look at each area of the websites appearance. For now, the nav bar is in place, time to move on.