12.10.2.2
Forgot to save my work for a while so here's a commit. I've made some good progress from what I wanted at the start. On mobile, a simple single column layout and larger resolutions splitting the page into 2 and letting the text scroll while the image stays fixed on the left.  There's a lot of tweaking to be done but I think the core of what I want for the layout is in place. It's now about the small details(padding, margins, font styles etc)