V12.10.4.6 - Removed Loading Bar, Kept loading number
The bar to indicate progress when swiping between articles wasnt looking great so i removed that and placed more focus on the number. Instead of change of loading bar colour to indicate the article has been loaded, the number changes to a "load complete" text. Still needs work but i already prefer this to the loading bar.