V12.10.4.10 - Adding box shadow to liteKnews container
I'm really quite happy with my work here. I'm beginning to put together a stronger visual identity for the website than I have before. The greyish colour im using is looking good as a secondary colour for the website and the box shadow i've added to the article container adds a nice effect when swiping in liteKnews