V12.10.7.7 - Add/Remove Bookmark using swipe/buttons
You can now add/remove bookmark using a swipe or the button or both. God, that took a while. The solution was to lift the state up from the oncard controls into the render of the card itself and update the render output based on the prop in the oncardcontrols.I think. Probably.