V12.10.7.2.2 - Added mark as hide function
Added a new way of hiding articles. Instead of pressing the button and the card hides instantly, the card now stays on screen until the page reloads or rerenders. I haven't fully integrated everything yet but the idea is to give the user more information and the opportunity to undo a card hide