V12.10.7.2.4 - Changes to hide functionality
Pressing the hide article button on a card will now not hide the card instantly.  Instead, you need to reload the page for these cards to hide .I am going to add a button to let users force reload the page but I think this is a better solution that instantly hiding articles as it gives you greater control over what you want to see.