12.10.7.7.4 - Adding a Page Layout button
I've added the start of a button that will let you change the page layout in addition to the look of the cards. This is in response to looking for a way around the stretching of cards that current occurs with the "freeform" layout. Offering more rigid  layouts gives users the option if they want it of something that might appeal to them. I'm not finished and there's something issues to be worked out, but i am happy with the plan and will continue working on the task.