12.10.7.7.5 - New Layouts in place
I have the new layouts in place. I've chosen 1x, 2x,3x and 4x columns layouts as the new layouts. The current cards work mostly well but for all of them the compact, thick and big squares card layouts need to be adjusted to fit correctly. The other card layouts are working as is.