12.10.7.12.6 - Working on Tags Author page
Having trouble getting the tags author page to work without first going to an article page. It's driving me up the wall but im at a point where i feel like im almost at a solution so im saving here. This is as good as I've been able to come up with so far.