12.10.7.13.7 - Finished Options Menu
All of the buttons are in place, the website is defaulting as I want it and the styling has been complete. I haven't detected any further issues with the rest of the project but I will keep an eye out for any issues. Other than that I now consider the options menu to be finished.