V12.10.8.5 - Code Cleanup
With the new scroll check code, the old scroll check code can be removed along with a whole bunch of other stuff that only existed to allow that code to function. There's a lot of stuff to be cleaned up so this might take a while. I started with the scroll check components themselves and moved onto the leftover articles name.